JavaScript (v3): libs - Move Prompter.logSeparator to Logger.logSeparator.
