+++ javascriptv3/example_code/sagemaker/tests/wkflw-sagemaker-geospatial-pipeline.integration.test.js
@@ -5,7 +5,7 @@
 
 import { describe, test } from "vitest";
 import { SageMakerPipelinesWkflw } from "../scenarios/wkflw-sagemaker-geospatial-pipeline/SageMakerPipelinesWkflw.js";
-import { Logger } from "libs/slow-logger.js";
+import { Logger } from "libs/logger.js";
 import { IAMClient } from "@aws-sdk/client-iam";
 import { SageMakerClient } from "@aws-sdk/client-sagemaker";
 import { S3Client } from "@aws-sdk/client-s3";
